new workflow for atomic publish

+++ plugins/pulp_rpm/plugins/distributors/yum/publish.py
@@ -13,6 +13,7 @@
 
 import os
 import shutil
+import time
 import traceback
 from gettext import gettext as _
 from pprint import pformat
@@ -47,6 +48,9 @@ from .reporting import (
     SUCCESSES, FAILURES, ERROR_DETAILS, PUBLISH_REPORT_KEYWORDS,
     new_progress_report, initialize_progress_sub_report, build_final_report)
 
+# FIXME: remove after merging of CLI refactorization
+from .reporting import new_progress_sub_report
+
 # -- constants -----------------------------------------------------------------
 
 _LOG = util.getLogger(__name__)
@@ -80,11 +84,15 @@ class Publisher(object):
         self.config = config
 
         self.progress_report = new_progress_report()
+        # FIXME: remove after merging of CLI refactorization
+        self.progress_report['publish_to_master'] = new_progress_sub_report()
         self.canceled = False
 
         self.package_dir = None
         self.repomd_file_context = None
         self.package_context = None
+
+        self.timestamp = str(time.time())
 
     @property
     def skip_list(self):
@@ -124,10 +132,14 @@ class Publisher(object):
                 PublishCompsStep(self).process()
                 PublishMetadataStep(self).process()
 
+            PublishToMasterRepository(self).process()
             PublishOverHttpStep(self).process()
             PublishOverHttpsStep(self).process()
 
             self._clear_directory(self.repo.working_dir)
+            self._clear_directory(configuration.get_master_publish_dir(self.repo),
+                                  skip_list=[self.timestamp])
+
         except Exception, e:
             # do bug log the details as the returned report has the details.
             _LOG.error(e)
@@ -168,12 +180,14 @@ class Publisher(object):
     # -- cleanup ---------------------------------------------------------------
 
     @staticmethod
-    def _clear_directory(path):
+    def _clear_directory(path, skip_list=()):
         """
         Clear out the contents of the given directory.
 
         :param path: path of the directory to clear out
         :type  path: str
+        :param skip_list: list of files or directories to not remove
+        :type  skip_list: list or tuple
         """
         _LOG.debug('Clearing out directory: %s' % path)
 
@@ -181,9 +195,15 @@ class Publisher(object):
             return
 
         for entry in os.listdir(path):
+
+            if entry in skip_list:
+                continue
+
             entry_path = os.path.join(path, entry)
+
             if os.path.isdir(entry_path):
                 shutil.rmtree(entry_path, ignore_errors=True)
+
             elif os.path.isfile(entry_path):
                 os.unlink(entry_path)
 
@@ -192,7 +212,7 @@ class PublishStep(object):
 
     def __init__(self, parent, step_id, unit_type=None):
         """
-        Set the default parent, step_id and unit_type for the the publsh step
+        Set the default parent, step_id and unit_type for the the publish step
         the unit_type defaults to none since some steps are not used for processing units.
 
         :param parent: The parent publisher that contains teh conduit, repo & cancel status
@@ -381,7 +401,8 @@ class PublishStep(object):
         :param step: step to initialize a progress sub-report for
         :type  step: str
         """
-        assert step in PUBLISH_STEPS
+        # FIXME: revert after merging of CLI refactorization
+        assert step in PUBLISH_STEPS or step == 'publish_to_master'
 
         initialize_progress_sub_report(self.parent.progress_report[step])
 
@@ -395,7 +416,8 @@ class PublishStep(object):
         :param report_details: keyword argument updates to the current step's
                                progress sub-report (if any)
         """
-        assert step in PUBLISH_STEPS
+        # FIXME: revert after merging of CLI refactorization
+        assert step in PUBLISH_STEPS or step == 'publish_to_master'
         assert set(report_details).issubset(set(PUBLISH_REPORT_KEYWORDS))
 
         self.parent.progress_report[step].update(report_details)
@@ -412,7 +434,8 @@ class PublishStep(object):
         :param tb: traceback instance (if any)
         :type  tb: Traceback or None
         """
-        assert step in PUBLISH_STEPS
+        # FIXME: revert after merging of CLI refactorization
+        assert step in PUBLISH_STEPS or step == 'publish_to_master'
 
         self.parent.progress_report[step][FAILURES] += 1
 
@@ -759,6 +782,27 @@ class PublishDistributionStep(PublishStep):
             self._create_symlink("./", packages_symlink_path)
 
 
+class PublishToMasterRepository(PublishStep):
+
+    def __init__(self, parent, step='publish_to_master'):
+        # FIXME: move 'step' to the constants after merging of the CLI refactorization
+        super(PublishToMasterRepository, self).__init__(parent, step, None)
+
+    def is_skipped(self):
+        return False
+
+    def _get_total(self, id_list=None):
+        return 1
+
+    def get_unit_generator(self):
+        return [configuration.get_master_publish_dir(self.parent.repo)]
+
+    def process_unit(self, master_publish_dir):
+        master_repo_directory = os.path.join(master_publish_dir, self.parent.timestamp)
+        _LOG.debug('Copying tree from %s to %s' % (self.parent.repo.working_dir, master_repo_directory))
+        shutil.copytree(self.parent.repo.working_dir, master_repo_directory, symlinks=True)
+
+
 class PublishOverHttpStep(PublishStep):
     """
     Publish http repo directory if configured
@@ -782,30 +826,39 @@ class PublishOverHttpStep(PublishStep):
         """
         Return a fake unit so that the process_unit method will be called
         """
-        return ['http']
+        return [configuration.get_http_publish_dir(self.parent.config)]
 
-    def process_unit(self, unit):
-        root_http_publish_dir = configuration.get_http_publish_dir(self.parent.config)
-        self._publish_repo_dir(root_http_publish_dir)
-
-    def _publish_repo_dir(self, publish_dir):
+    def process_unit(self, root_publish_dir):
         """
-        Publish a directory from the repo to a target directory.  This is split out to more easily
-        support subclassing for HTTPS support
+        Publish a directory from the repo to a target directory.
         """
-        repo_relative_dir = configuration.get_repo_relative_path(self.parent.repo,
-                                                                 self.parent.config)
-        repo_http_publish_dir = os.path.join(publish_dir, repo_relative_dir)
 
-        if os.path.exists(repo_http_publish_dir):
-            _LOG.debug('Removing old published directory: %s' % repo_http_publish_dir)
-            shutil.rmtree(repo_http_publish_dir)
+        # Find the location of the master repository tree structure
+        master_publish_dir = os.path.join(configuration.get_master_publish_dir(self.parent.repo),
+                                          self.parent.timestamp)
 
-        _LOG.debug('Copying tree from %s to %s' % (self.parent.repo.working_dir,
-                                                   repo_http_publish_dir))
-        shutil.copytree(self.parent.repo.working_dir, repo_http_publish_dir, symlinks=True)
+        # Find the location of the published repository tree structure
+        repo_relative_dir = configuration.get_repo_relative_path(self.parent.repo, self.parent.config)
+        repo_publish_dir = os.path.join(root_publish_dir, repo_relative_dir)
 
-        util.generate_listing_files(publish_dir, repo_http_publish_dir)
+        # Create the parent directory of the published repository tree, if needed
+        repo_publish_dir_parent = os.path.dirname(repo_publish_dir)
+        if not os.path.exists(repo_publish_dir_parent):
+            os.makedirs(repo_publish_dir_parent, 0750)
+
+        # Create a temporary symlink in the parent of the published directory tree
+        tmp_link_name = os.path.join(repo_publish_dir_parent, self.parent.timestamp)
+        os.symlink(master_publish_dir, tmp_link_name)
+
+        # Rename the symlink to the official published repository directory name.
+        # This has two desirable effects:
+        # 1. it will overwrite an existing link, if it's there
+        # 2. the operation is atomic, instantly changing the published directory
+        # NOTE: it's not easy (possible?) to directly edit the target of a symlink
+        os.rename(tmp_link_name, repo_publish_dir)
+
+        # (Re)generate the listing files
+        util.generate_listing_files(root_publish_dir, repo_publish_dir)
 
 
 class PublishOverHttpsStep(PublishOverHttpStep):
@@ -821,10 +874,5 @@ class PublishOverHttpsStep(PublishOverHttpStep):
         """
         return not self.parent.config.get('https')
 
-    def process_unit(self, unit):
-        """
-        Publish the files to the https directory & secure the directory
-        """
-        publish_dir = configuration.get_https_publish_dir(self.parent.config)
-        self._publish_repo_dir(publish_dir)
-        # TODO Configure certificate auth
+    def get_unit_generator(self):
+        return [configuration.get_https_publish_dir(self.parent.config)]
